Refactor Associations for bidirectional navigability
Type1/Type2 relationships are now stored in the Associations class. This will be converted to a list that will be updated for every type, and then output at the end of the run. This will allow us to implement [A]<->[B]. In addition, we'll be able to add a multiplicity flag to turn on explicit multiplicity for every association. Could also add names for the associations.